# [Early stage prototype] Optimize calls using vectorcall, 35% reduction in call overhead