… CMakeLists.txt
… include/pybind11/detail/internals.h
… include/pybind11/detail/python_compat.h
… include/pybind11/detail/vendor/optional.h
… include/pybind11/functional.h
… include/pybind11/pybind11.h
… include/pybind11/pytypes.h
… pybench/Makefile
… pybench/bench.py
… pybench/test_module.cc
… pybench/test_module_raw.cc
… tests/constructor_stats.h
… tests/extra_python_package/test_files.py
… tests/test_call_policies.cpp
… tests/test_call_policies.py
… tests/test_callbacks.py
… tests/test_class.cpp
… tests/test_class.py
… tests/test_factory_constructors.cpp
… tests/test_factory_constructors.py
… tests/test_kwargs_and_defaults.py
… tests/test_methods_and_attributes.py
… tests/test_multiple_inheritance.py
… tests/test_pickling.py
… tests/test_sequences_and_iterators.cpp
… tests/test_sequences_and_iterators.py
… tools/setup_global.py.in
… tools/setup_main.py.in