@Lalaland 2022-10-18
@lqf96 I think you might find this interesting given your prior work on https://github.com/pybind/pybind11/issues/2810