@Skylion007 2022-10-18
Lalaland, to try shorten the commit log a bit for this PR, please do the following:
`pip install pre-commit
# cd to git project root
pre-commit install
`
That will run pre-commit locally before every commit.