@henryiii 2022-10-19
3.8 added vectorcall. That’s why nanobind requires 3.8+.