@rwgk 2022-12-11
> old pybind11: 1.5625
> this PR: 6.1168060302734375

What exactly is 4 times bigger?

Sorry I don't have a lot of experience with micro benchmarks and binary size comparison.