@rwgk 2022-12-11
> > What exactly is 4 times bigger?
> 
> The .so files generated when the extensions are complied.

Wow ... I don't think I've ever seen such an explosion of binary sizes.

> I just need to create non inlined helper functions in several places.

Hm ... I did measurements back in Aug 2021:

https://docs.google.com/spreadsheets/d/1bkg1HtmU1AjotPTXCCLSbTBGARRet8W91noh1DnKOHU/edit#gid=0

noinline hardly made a difference, but maybe your situation is somehow special?