@Skylion007 2022-12-17
> > noinline hardly made a difference, but maybe your situation is somehow special?
> 
> The situation here is special because of the templates involved. So pulling an even small amount of code out of the template (and forcing it out of the template with non-inline) can have huge savings because of the hundreds of copies.
> 
> > Wow ... I don't think I've ever seen such an explosion of binary sizes.
> 
> Keep in mind that these microbenchmarks are sorta contrived. In order to make the sizes / compile times more significant the same code is copied 719 times.

This was also the case in the PR I linked here: https://github.com/pybind/pybind11/pull/2013/ they were eventually able to resolve it.