@henryiii 2024-06-23
Python 3.6 now dropped! Planning on dropping 3.7 immediately after the next release, too. That should make this much easier.